Replace Recover() with Run()
The behavior of Recover() was more complex than it needed to be, since
it would decide whether to run defers based on whether the program was
panicking. Instead of this, defers should always run.

This allows programs to be written in the form:

    func main() {
        defer defers.Run()
        if err := run(); err != nil {
            defers.Exit(1)
        }
    }

    func run() error {
        // Do something...
    }

And avoid the program exiting for any reason (besides a SIGKILL) without
running defers.